requirements.txt
flask>=2.0.0
flask-sqlalchemy>=3.0.0
pydantic>=2.0.0
pytest>=7.0.0
pytest-flask>=1.2.0
python-dotenv>=1.0.0  # Para carregar variáveis de ambiente